# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Auditions API",
    "description": "API Documentation for Movie Auditions Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Audition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "movie_id": {
            "type": "integer",
            "example": 1
          },
          "role": {
            "type": "string",
            "example": "Lead Actor"
          },
          "applicant_name": {
            "type": "string",
            "example": "John Doe"
          },
          "uploaded_videos": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "http://example.com/storage/audition_videos/video.mp4"
            }
          },
          "old_video_backups": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "http://example.com/storage/audition_videos/old_video.mp4"
            }
          },
          "notes": {
            "type": "string",
            "example": "Experienced actor with 5 years in the industry"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "viewed", "shortlisted", "rejected"],
            "example": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-04T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-04T12:00:00Z"
          }
        }
      },
...
    }
  },
  "paths": {
    "/auditions": {
      "get": {
        "tags": ["Auditions"],
        "summary": "Get list of auditions",
        "description": "Returns list of auditions for the authenticated user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "movie_title",
            "in": "query",
            "description": "Filter by movie title",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "Filter by role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort by column",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["created_at", "movie_title", "role", "status"]
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "current_page": {
                          "type": "integer"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Audition"
                          }
                        },
                        "first_page_url": {
                          "type": "string"
                        },
                        "from": {
                          "type": "integer"
                        },
                        "last_page": {
                          "type": "integer"
                        },
                        "last_page_url": {
                          "type": "string"
                        },
                        "links": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "url": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "active": {
                                "type": "boolean"
                              }
                            }
                          }
                        },
                        "next_page_url": {
                          "type": "string"
                        },
                        "path": {
                          "type": "string"
                        },
                        "per_page": {
                          "type": "integer"
                        },
                        "prev_page_url": {
                          "type": "string"
                        },
                        "to": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Auditions retrieved successfully."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Auditions"],
        "summary": "Store a new audition",
        "description": "Creates a new audition for the authenticated user",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "movie_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "role": {
                    "type": "string",
                    "example": "Lead Actor"
                  },
                  "applicant_name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "notes": {
                    "type": "string",
                    "example": "Experienced actor with 5 years in the industry"
                  },
                  "uploaded_videos": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Audition"
                    },
                    "message": {
                      "type": "string",
                      "example": "Audition created successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/auditions/{id}": {
      "get": {
        "tags": ["Auditions"],
        "summary": "Get audition information",
        "description": "Returns audition data for the authenticated user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Audition id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Audition"
                    },
                    "message": {
                      "type": "string",
                      "example": "Audition retrieved successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Audition not found"
          }
        }
      },
      "put": {
        "tags": ["Auditions"],
        "summary": "Update existing audition",
        "description": "Updates an audition for the authenticated user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Audition id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "example": "Lead Actor"
                  },
                  "applicant_name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "notes": {
                    "type": "string",
                    "example": "Updated notes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Audition"
                    },
                    "message": {
                      "type": "string",
                      "example": "Audition updated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Audition not found"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "delete": {
        "tags": ["Auditions"],
        "summary": "Delete audition",
        "description": "Deletes an audition for the authenticated user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Audition id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Audition deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Audition not found"
          }
        }
      }
    },
...
  }
}